ai-transcription/application/requirements.txt
PyYAML==6.0
psutil==5.9.5
pydub==0.25.1
watchdog==2.1.9
torch==2.0.0
numpy==1.24.3
python-telegram-bot==20.3
openai-whisper==20240930

# Packages for job queue and dashboard:
redis==4.6.0
rq==1.10.1
rq-dashboard==0.5.2
Flask==2.2.5
